refactor: :recycle: move steps dos tours para arquivo separado para evitar repetição
